[SPARK-47303][CORE][TESTS] Restructure MasterSuite
### What changes were proposed in this pull request?

This PR proposes to split `MasterSuite` as below:

```
master
├── MasterSuite.scala
├── MasterSuiteBase.scala
├── MasterWorkerUISuite.scala
├── RecoverySuite.scala
├── ResourceProfilesSuite.scala
└── WorkerSelectionSuite.scala
```

In addition, I reduced the memory and core specification in the tests below:

- Reduced the local cluster's work memory at `WorkerDecommissionExtendedSuite`
- Reduced memory at resource profiles at `ResourceProfilesSuite` and `ResourceProfilesMaxCoresSuite`
- Reduced the number of workers and cores at `ResourceProfilesMaxCoresSuite`

### Why are the changes needed?

MasterSuite takes too much resources, and it makes some tests fail with OOM with Mac OS: https://github.com/apache/spark/actions/runs/8162270810

### Does this PR introduce _any_ user-facing change?

No, test-only.

### How was this patch tested?

https://github.com/HyukjinKwon/spark/actions/runs/8170727017/job/22337438942

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45366 from HyukjinKwon/test-macos.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
